@@ -3,7 +3,7 @@
   "description": "Collaborative crafting made easy",
   "homepage": "https://ffxivteamcraft.com",
   "author": "FFXIV Teamcraft",
-  "version": "8.6.8",
+  "version": "9.0.6",
   "license": "MIT",
   "main": "dist/apps/electron/src/main.js",
   "build": {
@@ -118,11 +118,11 @@
   },
   "private": true,
   "dependencies": {
-    "@ffxiv-teamcraft/pcap-ffxiv": "^0.9.4",
+    "@ffxiv-teamcraft/pcap-ffxiv": "0.9.5",
     "@xivapi/nodestone": "^0.2.0",
     "buffer-reader": "^0.1.0",
     "electron-is-dev": "^1.1.0",
-    "electron-log": "^3.0.6",
+    "electron-log": "^4.4.1",
     "electron-store": "^5.2.0",
     "fs-extra": "^7.0.1",
     "hash-files": "^1.1.1",
@@ -218,9 +218,9 @@
     "deep-diff": "^1.0.2",
     "del-cli": "^1.1.0",
     "dotenv": "6.2.0",
-    "electron": "15.2.0",
+    "electron": "16.0.0",
     "electron-builder": "22.13.1",
-    "electron-builder-squirrel-windows": "22.14.5",
+    "electron-builder-squirrel-windows": "22.14.7",
     "enquirer": "^2.3.6",
     "express": "^4.17.1",
     "fast-json-patch": "^3.0.0-1",
