@@ -1,7 +1,7 @@
 import { LazyDataFacade } from '../../lazy-data/+state/lazy-data.facade';
 import { Observable, OperatorFunction } from 'rxjs';
 import { LazyDataKey, LazyDataWithExtracts } from '../../lazy-data/lazy-data-types';
-import { map, switchMap } from 'rxjs/operators';
+import { map, shareReplay, switchMap } from 'rxjs/operators';
 import { safeCombineLatest } from './safe-combine-latest';
 
 type TupleOf<T> = [T, ...T[]]
@@ -23,7 +23,8 @@ export function withLazyData<T extends any, K extends TupleOf<LazyDataKey>>(lazy
         return safeCombineLatest(
           entries.map(entry => lazyData.getEntry(entry))
         ).pipe(
-          map(entriesData => [data, ...entriesData] as [T, ...MapTuple<K>])
+          map(entriesData => [data, ...entriesData] as [T, ...MapTuple<K>]),
+          shareReplay(1)
         );
       })
     );
