@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Character } from '@xivapi/angular-client';
 import { Observable, of } from 'rxjs';
 import { LodestoneService } from '../../../core/api/lodestone.service';
-import { catchError, filter, map, shareReplay, startWith, tap } from 'rxjs/operators';
+import { catchError, filter, map, shareReplay, startWith } from 'rxjs/operators';
 import { TeamcraftUser } from '../../../model/user/teamcraft-user';
 import { UserService } from '../../../core/database/user.service';
 
