@@ -1,572 +1,572 @@
 export const aetheryteNames = {
-  '2': {
-    'en': 'Gridania Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ（冒険者・木工師ギルド/双蛇党）',
-    'fr': 'Place de l\'Éthérite (Guilde des aventuriers/menuisiers/Deux Vipères)'
-  },
-  '8': {
-    'en': 'Limsa Lominsa Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '9': {
-    'en': 'Ul\'dah Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ（格闘士ギルド/不滅隊）',
-    'fr': 'Place de l\'Éthérite (Guilde des pugilistes/Immortels)'
-  },
-  '25': {
-    'en': 'Archers\' Guild',
-    'de': 'Waldläufergilde',
-    'ja': '弓術士ギルド前',
-    'fr': 'Guilde des archers'
-  },
-  '26': {
-    'en': 'Leatherworkers\' Guild',
-    'de': 'Gerbergilde (Markt)',
-    'ja': '革細工師ギルド前（マーケット）',
-    'fr': 'Guilde des tanneurs (Marché)'
-  },
-  '27': {
-    'en': 'Lancers\' Guild',
-    'de': 'Pikeniergilde',
-    'ja': '槍術士ギルド前',
-    'fr': 'Guilde des maîtres d\'hast'
-  },
-  '28': {
-    'en': 'Conjurers\' Guild',
-    'de': 'Druidengilde',
-    'ja': '幻術士ギルド前',
-    'fr': 'Guilde des élémentalistes'
-  },
-  '29': {
-    'en': 'Botanists\' Guild',
-    'de': 'Gärtnergilde',
-    'ja': '園芸師ギルド前',
-    'fr': 'Guilde des botanistes'
-  },
-  '30': {
-    'en': 'Mih Khetto\'s Amphitheatre',
-    'de': 'Mih Khettos Amphitheater',
-    'ja': 'ミィ・ケット野外音楽堂',
-    'fr': 'Amphithéâtre de Mih Khetto'
-  },
-  '31': {
-    'en': 'Blue Badger Gate (Central Shroud)',
-    'de': 'Blaudachs-Pforte (zum Tiefen Wald / Ost)',
-    'ja': '青狢門（中央森林東方面）',
-    'fr': 'Forêt centrale est via la porte du Blaireau bleu'
-  },
-  '32': {
-    'en': 'Yellow Serpent Gate (North Shroud)',
-    'de': 'Gelbschlangen-Pforte (zum Nordwald)',
-    'ja': '黄蛇門（北部森林方面）',
-    'fr': 'Forêt du nord via la porte de la Vipère jaune'
-  },
-  '33': {
-    'en': 'Adventurers\' Guild',
-    'de': 'Abenteurergilde',
-    'ja': '冒険者ギルド前',
-    'fr': 'Guilde des aventuriers'
-  },
-  '34': {
-    'en': 'Thaumaturges\' Guild',
-    'de': 'Thaumaturgengilde',
-    'ja': '呪術士ギルド前',
-    'fr': 'Guilde des occultistes'
-  },
-  '35': {
-    'en': 'Gladiators\' Guild',
-    'de': 'Gladiatorengilde (Kolosseum)',
-    'ja': '剣術士ギルド前（コロセウム）',
-    'fr': 'Guilde des gladiateurs (Colisée)'
-  },
-  '36': {
-    'en': 'Miners\' Guild',
-    'de': 'Minenarbeitergilde (Himmlische Arena)',
-    'ja': '採掘師ギルド前（ブルースカイ）',
-    'fr': 'Guilde des mineurs (Ciel azur)'
-  },
-  '37': {
-    'en': 'Alchemists\' Guild',
-    'de': 'Alchemistengilde',
-    'ja': '錬金術師ギルド前',
-    'fr': 'Guilde des alchimistes'
-  },
-  '38': {
-    'en': 'Gate of the Sultana (Western Thanalan)',
-    'de': 'Neues Sultana-Tor (Westliches Thanalan)',
-    'ja': 'ナナモ新門（西ザナラーン方面）',
-    'fr': 'Thanalan occidental via la porte de Nanamo'
-  },
-  '39': {
-    'en': 'Gate of Nald (Central Thanalan)',
-    'de': 'Tor des Nald (Zentrales Thanalan)',
-    'ja': 'ナル大門（中央ザナラーン北方面）',
-    'fr': 'Thanalan central nord via la porte de Nald'
-  },
-  '40': {
-    'en': 'Gate of Thal (Central Thanalan)',
-    'de': 'Tor des Thal (Zentrales Thanalan)',
-    'ja': 'ザル大門（中央ザナラーン南方面）',
-    'fr': 'Thanalan central sud via la porte de Thal'
-  },
-  '41': {
-    'en': 'The Aftcastle',
-    'de': 'Achterburg',
-    'ja': '冒険者ギルド前（鍛冶師・甲冑師ギルド/黒渦団）',
-    'fr': 'Guilde des aventuriers (Guildes des armuriers & forgerons/Maelstrom)'
-  },
-  '42': {
-    'en': 'Culinarians\' Guild',
-    'de': 'Gourmetgilde (Restaurant Bismarck)',
-    'ja': '調理師ギルド前（レストラン「ビスマルク」）',
-    'fr': 'Guilde des cuisiniers (Restaurant Le Bismarck)'
-  },
-  '43': {
-    'en': 'Arcanists\' Guild',
-    'de': 'Hermetikergilde (Fähranleger)',
-    'ja': '巴術士ギルド前（フェリードック）',
-    'fr': 'Guilde des arcanistes (Embarcadère)'
-  },
-  '44': {
-    'en': 'Fishermen\'s Guild',
-    'de': 'Fischergilde (Schurkengilde)',
-    'ja': '漁師ギルド前（双剣士ギルド）',
-    'fr': 'L\'Escale du pêcheur (Guildes des pêcheurs & surineurs)'
-  },
-  '45': {
-    'en': 'Zephyr Gate (Middle La Noscea)',
-    'de': 'Zephyr-Tor (Zentrales La Noscea)',
-    'ja': 'ゼファー陸門（中央ラノシア方面）',
-    'fr': 'Noscea centrale via la porte du Zéphyr'
-  },
-  '46': {
-    'en': 'Tempest Gate (Lower La Noscea)',
-    'de': 'Sturmtor (Unteres La Noscea)',
-    'ja': 'テンペスト陸門（低地ラノシア方面）',
-    'fr': 'Basse-Noscea via la porte de la Tempête'
-  },
-  '47': {
-    'en': 'Weavers\' Guild',
-    'de': 'Webergilde',
-    'ja': '裁縫師ギルド前',
-    'fr': 'Guilde des couturiers'
-  },
-  '48': {
-    'en': 'Marauders\' Guild',
-    'de': 'Marodeurgilde (Korallenturm)',
-    'ja': '斧術士ギルド前（コーラルタワー）',
-    'fr': 'Guilde des maraudeurs (Tour de Corail)'
-  },
-  '49': {
-    'en': 'Hawkers\' Alley',
-    'de': 'Kaufmannsplatz',
-    'ja': 'マーケット（国際街広場）',
-    'fr': 'Marché (La Camelotade)'
-  },
-  '50': {
-    'en': 'Goldsmiths\' Guild',
-    'de': 'Goldschmiedegilde',
-    'ja': '彫金師ギルド前',
-    'fr': 'Guilde des orfèvres'
-  },
-  '51': {
-    'en': 'The Chamber of Rule',
-    'de': 'Kammer der Regentschaft',
-    'ja': 'ウルダハ王政庁',
-    'fr': 'Hall de la Loi'
-  },
-  '54': {
-    'en': 'White Wolf Gate (Central Shroud)',
-    'de': 'Weißwolf-Pforte (zum Tiefen Wald / West)',
-    'ja': '白狼門（中央森林西方面）',
-    'fr': 'Forêt centrale ouest via la porte du Loup blanc'
-  },
-  '62': {
-    'en': 'Gold Saucer Aetheryte Plaza',
-    'de': 'Gold Saucer-Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite du Gold Saucer'
-  },
-  '63': {
-    'en': 'Entrance & Card Squares',
-    'de': 'Foyer und Kartenspiel-Salon',
-    'ja': 'エントランススクウェア（カードスクウェア）',
-    'fr': 'Hall d\'accueil (Square des duellistes)'
-  },
-  '64': {
-    'en': 'Wonder Square East',
-    'de': 'Platz der Wunder (Ost)',
-    'ja': 'ワンダースクウェア東（ドマ式麻雀）',
-    'fr': 'Square des merveilles est'
-  },
-  '65': {
-    'en': 'Wonder Square West',
-    'de': 'Platz der Wunder (West)',
-    'ja': 'ワンダースクウェア西',
-    'fr': 'Square des merveilles ouest'
-  },
-  '66': {
-    'en': 'Event Square',
-    'de': 'Varieté-Saal',
-    'ja': 'イベントスクウェア',
-    'fr': 'Square des spectacles'
-  },
-  '67': {
-    'en': 'Cactpot Board',
-    'de': 'Glückskaktor-Schalter',
-    'ja': 'くじテンダーボード',
-    'fr': 'Comptoir du Méga Cactpot'
-  },
-  '68': {
-    'en': 'Round Square',
-    'de': 'Rondell',
-    'ja': 'ラウンドスクウェア',
-    'fr': 'Square du rond-point'
-  },
-  '69': {
-    'en': 'Chocobo Square',
-    'de': 'Chocobo-Platz',
-    'ja': 'チョコボスクウェア',
-    'fr': 'Square des chocobos'
-  },
-  '70': {
-    'en': 'Ishgard Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '75': {
-    'en': 'Idyllshire Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '80': {
-    'en': 'The Forgotten Knight',
-    'de': 'Der Vergessene Ritter',
-    'ja': '忘れられた騎士亭前（聖バルロアイアン広場）',
-    'fr': 'Le Chevalier oublié'
-  },
-  '81': {
-    'en': 'Skysteel Manufactory',
-    'de': 'Himmelsstahl-Werkstätte',
-    'ja': 'スカイスチール機工房前（聖レネット広場）',
-    'fr': 'Manufacture de Cielacier'
-  },
-  '82': {
-    'en': 'The Brume',
-    'de': 'Die Schwaden',
-    'ja': '雲霧街',
-    'fr': 'Brouillasse'
-  },
-  '83': {
-    'en': 'Athenaeum Astrologicum',
-    'de': 'Sankt Guenriol-Akademie für Astrologie (Luftschiff-Landeplatz)',
-    'ja': '聖ガンリオル占星院前（イシュガルド・ランディング）',
-    'fr': 'Collège d\'astromancie Saint-Guenriol (aérodrome d\'Ishgard)'
-  },
-  '84': {
-    'en': 'The Jeweled Crozier',
-    'de': 'Ferula-Zeile (Markt)',
-    'ja': 'マーケット（宝杖通り）',
-    'fr': 'Marché (Crosseron précieux)'
-  },
-  '85': {
-    'en': 'Saint Reymanaud\'s Cathedral',
-    'de': 'Sankt Reymanaud-Kathedrale (Sankt Endalim-Theologikum)',
-    'ja': '聖レマノー大聖堂（聖アンダリム神学院）',
-    'fr': 'Cathédrale Saint-Reymanaud (scolasticat Saint-Endalim)'
-  },
-  '86': {
-    'en': 'The Tribunal',
-    'de': 'Hochtribunal der Heiligen Inquisition',
-    'ja': '神聖裁判所',
-    'fr': 'Tribunal du Saint-Office'
-  },
-  '87': {
-    'en': 'The Last Vigil',
-    'de': 'Die Letzte Wacht (Anwesen der Fortemps)',
-    'ja': 'ラストヴィジル（フォルタン伯爵邸）',
-    'fr': 'Le Dernier Vigile (manoir des Fortemps)'
-  },
-  '88': {
-    'en': 'The Gates of Judgement (Coerthas Central Highlands)',
-    'de': 'Bogen des Würdigen',
-    'ja': '聖徒門（クルザス中央高地方面）',
-    'fr': 'Arc des Valeureux'
-  },
-  '89': {
-    'en': 'Minion Square',
-    'de': 'Chocobo-Platz (Halle der Trabanten)',
-    'ja': 'チョコボスクウェア（ミニオンスクウェア）',
-    'fr': 'Square des chocobos (square des mascottes)'
-  },
-  '90': {
-    'en': 'West Idyllshire',
-    'de': 'West-Frohehalde',
-    'ja': 'イディルシャイア西',
-    'fr': 'Idyllée ouest'
-  },
-  '91': {
-    'en': 'Prologue Gate (Western Hinterlands)',
-    'de': 'Vorworttor (Westliches Hinterland)',
-    'ja': '表紙門（低地ドラヴァニア西方面）',
-    'fr': 'Pontavant (Arrière-pays ouest)'
-  },
-  '92': {
-    'en': 'Epilogue Gate (Eastern Hinterlands)',
-    'de': 'Nachworttor (Östliches Hinterland)',
-    'ja': '裏表紙門（低地ドラヴァニア東方面）',
-    'fr': 'Pontarrière (Arrière-pays est)'
-  },
-  '93': {
-    'en': 'Airship Landing',
-    'de': 'Luftschiff-Landeplatz',
-    'ja': 'リムサ・ロミンサ・ランディング',
-    'fr': 'Aérodrome de Limsa Lominsa'
-  },
-  '94': {
-    'en': 'Airship Landing',
-    'de': 'Luftschiff-Landeplatz (Gridania)',
-    'ja': 'グリダニア・ランディング',
-    'fr': 'Aérodrome de Gridania'
-  },
-  '95': {
-    'en': 'Airship Landing',
-    'de': 'Luftschiff-Landeplatz',
-    'ja': 'ウルダハ・ランディング',
-    'fr': 'Aérodrome d\'Ul\'dah'
-  },
-  '104': {
-    'en': 'Rhalgr\'s Reach Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '111': {
-    'en': 'Kugane Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place Tenkonto'
-  },
-  '112': {
-    'en': 'Shiokaze Hostelry',
-    'de': 'Gasthaus Shiokaze',
-    'ja': '潮風亭前',
-    'fr': 'Shiokaze'
-  },
-  '113': {
-    'en': 'Pier #1',
-    'de': 'Pier 1',
-    'ja': '第一波止場',
-    'fr': 'Jetée n°1'
-  },
-  '114': {
-    'en': 'Thavnairian Consulate',
-    'de': 'Konsulat von Thavnair',
-    'ja': 'ラザハン大使館前',
-    'fr': 'Consulat de Radz-at-Han'
-  },
-  '115': {
-    'en': 'Kogane Dori Markets',
-    'de': 'Kogane-dori (Markt)',
-    'ja': 'マーケット（小金通り）',
-    'fr': 'Marché (Kogane-dori)'
-  },
-  '116': {
-    'en': 'Bokairo Inn',
-    'de': 'Badehaus Bokairo',
-    'ja': '望海楼',
-    'fr': 'Auberge du Bokairo'
-  },
-  '117': {
-    'en': 'The Ruby Bazaar',
-    'de': 'Ostaldenard-Handelshaus',
-    'ja': 'ウルダハ商館前',
-    'fr': 'Bazar de Rubis'
-  },
-  '118': {
-    'en': 'Sekiseigumi Barracks',
-    'de': 'Sekiseigumi-Quartier',
-    'ja': '赤誠組屯所南',
-    'fr': 'Caserne du Sekiseigumi'
-  },
-  '119': {
-    'en': 'Rakuza District',
-    'de': 'Rakuza-Bezirk',
-    'ja': '楽座街',
-    'fr': 'Quartier de Rakuza'
-  },
-  '120': {
-    'en': 'The Ruby Price',
-    'de': 'Rubin-Seefestung',
-    'ja': '紅玉海方面',
-    'fr': 'Mer de Rubis'
-  },
-  '121': {
-    'en': 'Western Rhalgr\'s Reach',
-    'de': 'Rhalgrs Wacht (Westteil)',
-    'ja': 'ラールガーズリーチ西',
-    'fr': 'L\'Étendue de Rhalgr - ouest'
-  },
-  '122': {
-    'en': 'Northeastern Rhalgr\'s Reach',
-    'de': 'Rhalgrs Wacht (Nordostteil)',
-    'ja': 'ラールガーズリーチ北東',
-    'fr': 'L\'Étendue de Rhalgr - nord-est'
-  },
-  '123': {
-    'en': 'Fringes Gate',
-    'de': 'Tor zum Grenzland',
-    'ja': 'ギラバニア辺境地帯方面',
-    'fr': 'Les Marges'
-  },
-  '124': {
-    'en': 'Peaks Gate',
-    'de': 'Tor zu den Zinnen',
-    'ja': 'ギラバニア山岳地帯方面',
-    'fr': 'Les Pics'
-  },
-  '125': {
-    'en': 'Sapphire Avenue Exchange',
-    'de': 'Saphirstraßenmarkt',
-    'ja': 'サファイアアベニュー国際市場（マーケット）',
-    'fr': 'Marchés de l\'avenue du Saphir'
-  },
-  '126': {
-    'en': 'Airship Landing',
-    'de': 'Luftschiff-Landeplatz',
-    'ja': 'クガネ・ランディング',
-    'fr': 'Aérodrome de Kugane'
-  },
-  '127': {
-    'en': 'Doman Enclave Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '129': {
-    'en': 'The Northern Enclave',
-    'de': 'Enklave Nord (Reisfelder und Einzig-Garten)',
-    'ja': 'ドマ町人地北（無二庭園・田園地）',
-    'fr': 'Quartier enclavé nord (jardins de Muniteien/rizières de Den\'en-chi)'
-  },
-  '130': {
-    'en': 'The Southern Enclave',
-    'de': 'Enklave Süd (Rissai-Schule)',
-    'ja': 'ドマ町人地南（栗斎塾）',
-    'fr': 'Quartier enclavé sud (école Rissai-juku)'
-  },
-  '131': {
-    'en': 'The One River',
-    'de': 'Goldbecken (Yanxia)',
-    'ja': '無二江流域（ヤンサ方面）',
-    'fr': 'Bassin rutilant (Yanxia)'
-  },
-  '133': {
-    'en': 'The Crystarium Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '134': {
-    'en': 'Eulmore Aetheryte Plaza',
-    'de': 'Ätherytenplatz',
-    'ja': 'エーテライト・プラザ',
-    'fr': 'Place de l\'Éthérite'
-  },
-  '135': {
-    'en': 'Southeast Derelicts',
-    'de': 'Wrackstadt (Südosten)',
-    'ja': '廃船街南東',
-    'fr': 'Sud-est de La Fange'
-  },
-  '149': {
-    'en': 'Musica Universalis Markets',
-    'de': 'Musica Universalis',
-    'ja': 'マーケット（ムジカ・ユニバーサリス）',
-    'fr': 'Marché (Musica Universalis)'
-  },
-  '150': {
-    'en': 'Temenos Rookery',
-    'de': 'Tierhof Temenos',
-    'ja': 'テメノスルカリー牧場',
-    'fr': 'Couvée de Temenos'
-  },
-  '151': {
-    'en': 'The Dossal Gate',
-    'de': 'Propylaeum (Okular)',
-    'ja': 'ドッサル大門前（星見の間）',
-    'fr': 'La Courtine'
-  },
-  '152': {
-    'en': 'The Pendants',
-    'de': 'Isomorph-Quartiere',
-    'ja': 'ペンダント居住館前',
-    'fr': 'Les Pendants'
-  },
-  '153': {
-    'en': 'The Amaro Launch',
-    'de': 'Amaro-Flugsteig',
-    'ja': 'アマロ桟橋前',
-    'fr': 'Jetée des amaros'
-  },
-  '154': {
-    'en': 'The Crystalline Mean',
-    'de': 'Manufaktur der Gemeinschaft',
-    'ja': 'ミーン工芸館前',
-    'fr': 'Halle du Juste'
-  },
-  '155': {
-    'en': 'The Cabinet of Curiosity',
-    'de': 'Kabinett der Kuriositäten',
-    'ja': '博物陳列館前',
-    'fr': 'Cabinet de curiosités'
-  },
-  '156': {
-    'en': 'Tessellation (Lakeland)',
-    'de': 'Mosaikbrücke (ins Seenland)',
-    'ja': 'テセレーション鉄橋（レイクランド方面）',
-    'fr': 'Le Pavage'
-  },
-  '157': {
-    'en': 'The Mainstay',
-    'de': 'Grundpfeiler der Ordnung',
-    'ja': 'メインステイ',
-    'fr': 'Ballast principal'
-  },
-  '158': {
-    'en': 'Nightsoil Pots',
-    'de': 'Gärten der Gestrigen Genüsse',
-    'ja': '腐臭の花園',
-    'fr': 'Les Goguenots fleuris'
-  },
-  '159': {
-    'en': 'The Glory Gate',
-    'de': 'Tor der Herrlichkeit',
-    'ja': '栄光の門',
-    'fr': 'Porte de la Gloire'
-  },
-  '160': {
-    'en': 'The Path to Glory (Kholusia)',
-    'de': 'Pfad zum Ruhm (Kholusia)',
-    'ja': '栄光の道（コルシア島方面）',
-    'fr': 'Marches de la Gloire (Kholusia)'
-  },
-  '162': {
-    'en': 'Ferry Docks',
-    'de': 'Fähranleger',
-    'ja': '船着場',
-    'fr': 'Embarcadère'
-  },
-  '163': {
-    'en': 'Gangos',
-    'de': 'Gangos',
-    'ja': 'ガンゴッシュ',
-    'fr': 'Gangos'
+  "2": {
+    "en": "Gridania Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ（冒険者・木工師ギルド/双蛇党）",
+    "fr": "Place de l'Éthérite (Guilde des aventuriers/menuisiers/Deux Vipères)"
+  },
+  "8": {
+    "en": "Limsa Lominsa Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "9": {
+    "en": "Ul'dah Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ（格闘士ギルド/不滅隊）",
+    "fr": "Place de l'Éthérite (Guilde des pugilistes/Immortels)"
+  },
+  "25": {
+    "en": "Archers' Guild",
+    "de": "Waldläufergilde",
+    "ja": "弓術士ギルド前",
+    "fr": "Guilde des archers"
+  },
+  "26": {
+    "en": "Leatherworkers' Guild",
+    "de": "Gerbergilde (Markt)",
+    "ja": "革細工師ギルド前（マーケット）",
+    "fr": "Guilde des tanneurs (Marché)"
+  },
+  "27": {
+    "en": "Lancers' Guild",
+    "de": "Pikeniergilde",
+    "ja": "槍術士ギルド前",
+    "fr": "Guilde des maîtres d'hast"
+  },
+  "28": {
+    "en": "Conjurers' Guild",
+    "de": "Druidengilde",
+    "ja": "幻術士ギルド前",
+    "fr": "Guilde des élémentalistes"
+  },
+  "29": {
+    "en": "Botanists' Guild",
+    "de": "Gärtnergilde",
+    "ja": "園芸師ギルド前",
+    "fr": "Guilde des botanistes"
+  },
+  "30": {
+    "en": "Mih Khetto's Amphitheatre",
+    "de": "Mih Khettos Amphitheater",
+    "ja": "ミィ・ケット野外音楽堂",
+    "fr": "Amphithéâtre de Mih Khetto"
+  },
+  "31": {
+    "en": "Blue Badger Gate (Central Shroud)",
+    "de": "Blaudachs-Pforte (zum Tiefen Wald / Ost)",
+    "ja": "青狢門（中央森林東方面）",
+    "fr": "Forêt centrale est via la porte du Blaireau bleu"
+  },
+  "32": {
+    "en": "Yellow Serpent Gate (North Shroud)",
+    "de": "Gelbschlangen-Pforte (zum Nordwald)",
+    "ja": "黄蛇門（北部森林方面）",
+    "fr": "Forêt du nord via la porte de la Vipère jaune"
+  },
+  "33": {
+    "en": "Adventurers' Guild",
+    "de": "Abenteurergilde",
+    "ja": "冒険者ギルド前",
+    "fr": "Guilde des aventuriers"
+  },
+  "34": {
+    "en": "Thaumaturges' Guild",
+    "de": "Thaumaturgengilde",
+    "ja": "呪術士ギルド前",
+    "fr": "Guilde des occultistes"
+  },
+  "35": {
+    "en": "Gladiators' Guild",
+    "de": "Gladiatorengilde (Kolosseum)",
+    "ja": "剣術士ギルド前（コロセウム）",
+    "fr": "Guilde des gladiateurs (Colisée)"
+  },
+  "36": {
+    "en": "Miners' Guild",
+    "de": "Minenarbeitergilde (Himmlische Arena)",
+    "ja": "採掘師ギルド前（ブルースカイ）",
+    "fr": "Guilde des mineurs (Ciel azur)"
+  },
+  "37": {
+    "en": "Alchemists' Guild",
+    "de": "Alchemistengilde",
+    "ja": "錬金術師ギルド前",
+    "fr": "Guilde des alchimistes"
+  },
+  "38": {
+    "en": "Gate of the Sultana (Western Thanalan)",
+    "de": "Neues Sultana-Tor (Westliches Thanalan)",
+    "ja": "ナナモ新門（西ザナラーン方面）",
+    "fr": "Thanalan occidental via la porte de Nanamo"
+  },
+  "39": {
+    "en": "Gate of Nald (Central Thanalan)",
+    "de": "Tor des Nald (Zentrales Thanalan)",
+    "ja": "ナル大門（中央ザナラーン北方面）",
+    "fr": "Thanalan central nord via la porte de Nald"
+  },
+  "40": {
+    "en": "Gate of Thal (Central Thanalan)",
+    "de": "Tor des Thal (Zentrales Thanalan)",
+    "ja": "ザル大門（中央ザナラーン南方面）",
+    "fr": "Thanalan central sud via la porte de Thal"
+  },
+  "41": {
+    "en": "The Aftcastle",
+    "de": "Achterburg",
+    "ja": "冒険者ギルド前（鍛冶師・甲冑師ギルド/黒渦団）",
+    "fr": "Guilde des aventuriers (Guildes des armuriers & forgerons/Maelstrom)"
+  },
+  "42": {
+    "en": "Culinarians' Guild",
+    "de": "Gourmetgilde (Restaurant Bismarck)",
+    "ja": "調理師ギルド前（レストラン「ビスマルク」）",
+    "fr": "Guilde des cuisiniers (Restaurant Le Bismarck)"
+  },
+  "43": {
+    "en": "Arcanists' Guild",
+    "de": "Hermetikergilde (Fähranleger)",
+    "ja": "巴術士ギルド前（フェリードック）",
+    "fr": "Guilde des arcanistes (Embarcadère)"
+  },
+  "44": {
+    "en": "Fishermen's Guild",
+    "de": "Fischergilde (Schurkengilde)",
+    "ja": "漁師ギルド前（双剣士ギルド）",
+    "fr": "L'Escale du pêcheur (Guildes des pêcheurs & surineurs)"
+  },
+  "45": {
+    "en": "Zephyr Gate (Middle La Noscea)",
+    "de": "Zephyr-Tor (Zentrales La Noscea)",
+    "ja": "ゼファー陸門（中央ラノシア方面）",
+    "fr": "Noscea centrale via la porte du Zéphyr"
+  },
+  "46": {
+    "en": "Tempest Gate (Lower La Noscea)",
+    "de": "Sturmtor (Unteres La Noscea)",
+    "ja": "テンペスト陸門（低地ラノシア方面）",
+    "fr": "Basse-Noscea via la porte de la Tempête"
+  },
+  "47": {
+    "en": "Weavers' Guild",
+    "de": "Webergilde",
+    "ja": "裁縫師ギルド前",
+    "fr": "Guilde des couturiers"
+  },
+  "48": {
+    "en": "Marauders' Guild",
+    "de": "Marodeurgilde (Korallenturm)",
+    "ja": "斧術士ギルド前（コーラルタワー）",
+    "fr": "Guilde des maraudeurs (Tour de Corail)"
+  },
+  "49": {
+    "en": "Hawkers' Alley",
+    "de": "Kaufmannsplatz",
+    "ja": "マーケット（国際街広場）",
+    "fr": "Marché (La Camelotade)"
+  },
+  "50": {
+    "en": "Goldsmiths' Guild",
+    "de": "Goldschmiedegilde",
+    "ja": "彫金師ギルド前",
+    "fr": "Guilde des orfèvres"
+  },
+  "51": {
+    "en": "The Chamber of Rule",
+    "de": "Kammer der Regentschaft",
+    "ja": "ウルダハ王政庁",
+    "fr": "Hall de la Loi"
+  },
+  "54": {
+    "en": "White Wolf Gate (Central Shroud)",
+    "de": "Weißwolf-Pforte (zum Tiefen Wald / West)",
+    "ja": "白狼門（中央森林西方面）",
+    "fr": "Forêt centrale ouest via la porte du Loup blanc"
+  },
+  "62": {
+    "en": "Gold Saucer Aetheryte Plaza",
+    "de": "Gold Saucer-Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite du Gold Saucer"
+  },
+  "63": {
+    "en": "Entrance & Card Squares",
+    "de": "Foyer und Kartenspiel-Salon",
+    "ja": "エントランススクウェア（カードスクウェア）",
+    "fr": "Hall d'accueil (Square des duellistes)"
+  },
+  "64": {
+    "en": "Wonder Square East",
+    "de": "Platz der Wunder (Ost)",
+    "ja": "ワンダースクウェア東（ドマ式麻雀）",
+    "fr": "Square des merveilles est"
+  },
+  "65": {
+    "en": "Wonder Square West",
+    "de": "Platz der Wunder (West)",
+    "ja": "ワンダースクウェア西",
+    "fr": "Square des merveilles ouest"
+  },
+  "66": {
+    "en": "Event Square",
+    "de": "Varieté-Saal",
+    "ja": "イベントスクウェア",
+    "fr": "Square des spectacles"
+  },
+  "67": {
+    "en": "Cactpot Board",
+    "de": "Glückskaktor-Schalter",
+    "ja": "くじテンダーボード",
+    "fr": "Comptoir du Méga Cactpot"
+  },
+  "68": {
+    "en": "Round Square",
+    "de": "Rondell",
+    "ja": "ラウンドスクウェア",
+    "fr": "Square du rond-point"
+  },
+  "69": {
+    "en": "Chocobo Square",
+    "de": "Chocobo-Platz",
+    "ja": "チョコボスクウェア",
+    "fr": "Square des chocobos"
+  },
+  "70": {
+    "en": "Ishgard Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "75": {
+    "en": "Idyllshire Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "80": {
+    "en": "The Forgotten Knight",
+    "de": "Der Vergessene Ritter",
+    "ja": "忘れられた騎士亭前（聖バルロアイアン広場）",
+    "fr": "Le Chevalier oublié"
+  },
+  "81": {
+    "en": "Skysteel Manufactory",
+    "de": "Himmelsstahl-Werkstätte",
+    "ja": "スカイスチール機工房前（聖レネット広場）",
+    "fr": "Manufacture de Cielacier"
+  },
+  "82": {
+    "en": "The Brume",
+    "de": "Die Schwaden",
+    "ja": "雲霧街",
+    "fr": "Brouillasse"
+  },
+  "83": {
+    "en": "Athenaeum Astrologicum",
+    "de": "Sankt Guenriol-Akademie für Astrologie (Luftschiff-Landeplatz)",
+    "ja": "聖ガンリオル占星院前（イシュガルド・ランディング）",
+    "fr": "Collège d'astromancie Saint-Guenriol (aérodrome d'Ishgard)"
+  },
+  "84": {
+    "en": "The Jeweled Crozier",
+    "de": "Ferula-Zeile (Markt)",
+    "ja": "マーケット（宝杖通り）",
+    "fr": "Marché (Crosseron précieux)"
+  },
+  "85": {
+    "en": "Saint Reymanaud's Cathedral",
+    "de": "Sankt Reymanaud-Kathedrale (Sankt Endalim-Theologikum)",
+    "ja": "聖レマノー大聖堂（聖アンダリム神学院）",
+    "fr": "Cathédrale Saint-Reymanaud (scolasticat Saint-Endalim)"
+  },
+  "86": {
+    "en": "The Tribunal",
+    "de": "Hochtribunal der Heiligen Inquisition",
+    "ja": "神聖裁判所",
+    "fr": "Tribunal du Saint-Office"
+  },
+  "87": {
+    "en": "The Last Vigil",
+    "de": "Die Letzte Wacht (Anwesen der Fortemps)",
+    "ja": "ラストヴィジル（フォルタン伯爵邸）",
+    "fr": "Le Dernier Vigile (manoir des Fortemps)"
+  },
+  "88": {
+    "en": "The Gates of Judgement (Coerthas Central Highlands)",
+    "de": "Bogen des Würdigen",
+    "ja": "聖徒門（クルザス中央高地方面）",
+    "fr": "Arc des Valeureux"
+  },
+  "89": {
+    "en": "Minion Square",
+    "de": "Chocobo-Platz (Halle der Trabanten)",
+    "ja": "チョコボスクウェア（ミニオンスクウェア）",
+    "fr": "Square des chocobos (square des mascottes)"
+  },
+  "90": {
+    "en": "West Idyllshire",
+    "de": "West-Frohehalde",
+    "ja": "イディルシャイア西",
+    "fr": "Idyllée ouest"
+  },
+  "91": {
+    "en": "Prologue Gate (Western Hinterlands)",
+    "de": "Vorworttor (Westliches Hinterland)",
+    "ja": "表紙門（低地ドラヴァニア西方面）",
+    "fr": "Pontavant (Arrière-pays ouest)"
+  },
+  "92": {
+    "en": "Epilogue Gate (Eastern Hinterlands)",
+    "de": "Nachworttor (Östliches Hinterland)",
+    "ja": "裏表紙門（低地ドラヴァニア東方面）",
+    "fr": "Pontarrière (Arrière-pays est)"
+  },
+  "93": {
+    "en": "Airship Landing",
+    "de": "Luftschiff-Landeplatz",
+    "ja": "リムサ・ロミンサ・ランディング",
+    "fr": "Aérodrome de Limsa Lominsa"
+  },
+  "94": {
+    "en": "Airship Landing",
+    "de": "Luftschiff-Landeplatz (Gridania)",
+    "ja": "グリダニア・ランディング",
+    "fr": "Aérodrome de Gridania"
+  },
+  "95": {
+    "en": "Airship Landing",
+    "de": "Luftschiff-Landeplatz",
+    "ja": "ウルダハ・ランディング",
+    "fr": "Aérodrome d'Ul'dah"
+  },
+  "104": {
+    "en": "Rhalgr's Reach Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "111": {
+    "en": "Kugane Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place Tenkonto"
+  },
+  "112": {
+    "en": "Shiokaze Hostelry",
+    "de": "Gasthaus Shiokaze",
+    "ja": "潮風亭前",
+    "fr": "Shiokaze"
+  },
+  "113": {
+    "en": "Pier #1",
+    "de": "Pier 1",
+    "ja": "第一波止場",
+    "fr": "Jetée n°1"
+  },
+  "114": {
+    "en": "Thavnairian Consulate",
+    "de": "Konsulat von Thavnair",
+    "ja": "ラザハン大使館前",
+    "fr": "Consulat de Radz-at-Han"
+  },
+  "115": {
+    "en": "Kogane Dori Markets",
+    "de": "Kogane-dori (Markt)",
+    "ja": "マーケット（小金通り）",
+    "fr": "Marché (Kogane-dori)"
+  },
+  "116": {
+    "en": "Bokairo Inn",
+    "de": "Badehaus Bokairo",
+    "ja": "望海楼",
+    "fr": "Auberge du Bokairo"
+  },
+  "117": {
+    "en": "The Ruby Bazaar",
+    "de": "Ostaldenard-Handelshaus",
+    "ja": "ウルダハ商館前",
+    "fr": "Bazar de Rubis"
+  },
+  "118": {
+    "en": "Sekiseigumi Barracks",
+    "de": "Sekiseigumi-Quartier",
+    "ja": "赤誠組屯所南",
+    "fr": "Caserne du Sekiseigumi"
+  },
+  "119": {
+    "en": "Rakuza District",
+    "de": "Rakuza-Bezirk",
+    "ja": "楽座街",
+    "fr": "Quartier de Rakuza"
+  },
+  "120": {
+    "en": "The Ruby Price",
+    "de": "Rubin-Seefestung",
+    "ja": "紅玉海方面",
+    "fr": "Mer de Rubis"
+  },
+  "121": {
+    "en": "Western Rhalgr's Reach",
+    "de": "Rhalgrs Wacht (Westteil)",
+    "ja": "ラールガーズリーチ西",
+    "fr": "L'Étendue de Rhalgr - ouest"
+  },
+  "122": {
+    "en": "Northeastern Rhalgr's Reach",
+    "de": "Rhalgrs Wacht (Nordostteil)",
+    "ja": "ラールガーズリーチ北東",
+    "fr": "L'Étendue de Rhalgr - nord-est"
+  },
+  "123": {
+    "en": "Fringes Gate",
+    "de": "Tor zum Grenzland",
+    "ja": "ギラバニア辺境地帯方面",
+    "fr": "Les Marges"
+  },
+  "124": {
+    "en": "Peaks Gate",
+    "de": "Tor zu den Zinnen",
+    "ja": "ギラバニア山岳地帯方面",
+    "fr": "Les Pics"
+  },
+  "125": {
+    "en": "Sapphire Avenue Exchange",
+    "de": "Saphirstraßenmarkt",
+    "ja": "サファイアアベニュー国際市場（マーケット）",
+    "fr": "Marchés de l'avenue du Saphir"
+  },
+  "126": {
+    "en": "Airship Landing",
+    "de": "Luftschiff-Landeplatz",
+    "ja": "クガネ・ランディング",
+    "fr": "Aérodrome de Kugane"
+  },
+  "127": {
+    "en": "Doman Enclave Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "129": {
+    "en": "The Northern Enclave",
+    "de": "Enklave Nord (Reisfelder und Einzig-Garten)",
+    "ja": "ドマ町人地北（無二庭園・田園地）",
+    "fr": "Quartier enclavé nord (jardins de Muniteien/rizières de Den'en-chi)"
+  },
+  "130": {
+    "en": "The Southern Enclave",
+    "de": "Enklave Süd (Rissai-Schule)",
+    "ja": "ドマ町人地南（栗斎塾）",
+    "fr": "Quartier enclavé sud (école Rissai-juku)"
+  },
+  "131": {
+    "en": "The One River",
+    "de": "Goldbecken (Yanxia)",
+    "ja": "無二江流域（ヤンサ方面）",
+    "fr": "Bassin rutilant (Yanxia)"
+  },
+  "133": {
+    "en": "The Crystarium Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "134": {
+    "en": "Eulmore Aetheryte Plaza",
+    "de": "Ätherytenplatz",
+    "ja": "エーテライト・プラザ",
+    "fr": "Place de l'Éthérite"
+  },
+  "135": {
+    "en": "Southeast Derelicts",
+    "de": "Wrackstadt (Südosten)",
+    "ja": "廃船街南東",
+    "fr": "Sud-est de La Fange"
+  },
+  "149": {
+    "en": "Musica Universalis Markets",
+    "de": "Musica Universalis",
+    "ja": "マーケット（ムジカ・ユニバーサリス）",
+    "fr": "Marché (Musica Universalis)"
+  },
+  "150": {
+    "en": "Temenos Rookery",
+    "de": "Tierhof Temenos",
+    "ja": "テメノスルカリー牧場",
+    "fr": "Couvée de Temenos"
+  },
+  "151": {
+    "en": "The Dossal Gate",
+    "de": "Propylaeum (Okular)",
+    "ja": "ドッサル大門前（星見の間）",
+    "fr": "La Courtine"
+  },
+  "152": {
+    "en": "The Pendants",
+    "de": "Isomorph-Quartiere",
+    "ja": "ペンダント居住館前",
+    "fr": "Les Pendants"
+  },
+  "153": {
+    "en": "The Amaro Launch",
+    "de": "Amaro-Flugsteig",
+    "ja": "アマロ桟橋前",
+    "fr": "Jetée des amaros"
+  },
+  "154": {
+    "en": "The Crystalline Mean",
+    "de": "Manufaktur der Gemeinschaft",
+    "ja": "ミーン工芸館前",
+    "fr": "Halle du Juste"
+  },
+  "155": {
+    "en": "The Cabinet of Curiosity",
+    "de": "Kabinett der Kuriositäten",
+    "ja": "博物陳列館前",
+    "fr": "Cabinet de curiosités"
+  },
+  "156": {
+    "en": "Tessellation (Lakeland)",
+    "de": "Mosaikbrücke (ins Seenland)",
+    "ja": "テセレーション鉄橋（レイクランド方面）",
+    "fr": "Le Pavage"
+  },
+  "157": {
+    "en": "The Mainstay",
+    "de": "Grundpfeiler der Ordnung",
+    "ja": "メインステイ",
+    "fr": "Ballast principal"
+  },
+  "158": {
+    "en": "Nightsoil Pots",
+    "de": "Gärten der Gestrigen Genüsse",
+    "ja": "腐臭の花園",
+    "fr": "Les Goguenots fleuris"
+  },
+  "159": {
+    "en": "The Glory Gate",
+    "de": "Tor der Herrlichkeit",
+    "ja": "栄光の門",
+    "fr": "Porte de la Gloire"
+  },
+  "160": {
+    "en": "The Path to Glory (Kholusia)",
+    "de": "Pfad zum Ruhm (Kholusia)",
+    "ja": "栄光の道（コルシア島方面）",
+    "fr": "Marches de la Gloire (Kholusia)"
+  },
+  "162": {
+    "en": "Ferry Docks",
+    "de": "Fähranleger",
+    "ja": "船着場",
+    "fr": "Embarcadère"
+  },
+  "163": {
+    "en": "Gangos",
+    "de": "Gangos",
+    "ja": "ガンゴッシュ",
+    "fr": "Gangos"
   }
 };
