@@ -1,4 +1,4 @@
-import { Aetheryte } from '../../core/data/aetheryte';
+import { LazyAetheryte } from '../../lazy-data/model/lazy-aetheryte';
 
 export interface MapData {
   id: number;
@@ -15,5 +15,5 @@ export interface MapData {
   placename_id: number;
   placename_sub_id: number;
   region_id: number;
-  aetherytes?: Aetheryte[];
+  aetherytes?: LazyAetheryte[];
 }
