@@ -88,6 +88,9 @@ export class ItemComponent extends TeamcraftPageComponent implements OnInit, OnD
       if (item.ItemSeries) {
         return this.lazyData.getRow('itemSeries', item.ItemSeries.ID).pipe(
           switchMap(itemSeries => {
+            if(itemSeries.items.length > 20){
+              return of(item);
+            }
             return safeCombineLatest(itemSeries.items.map(itemId => {
               return this.lazyData.getRow('itemSetBonuses', itemId, { bonuses: [], itemSeriesId: item.ItemSeries.ID }).pipe(
                 map(bonuses => {
@@ -97,11 +100,17 @@ export class ItemComponent extends TeamcraftPageComponent implements OnInit, OnD
                   };
                 })
               );
-            }));
+            })).pipe(
+              map(bonuses => {
+                return {
+                  ...item,
+                  bonuses
+                }
+              })
+            );
           })
         );
       }
-
       return of(item);
     }),
     shareReplay(1)
@@ -888,6 +897,7 @@ export class ItemComponent extends TeamcraftPageComponent implements OnInit, OnD
         })
       );
     }
+
     return res$;
   }
 
