@@ -31,7 +31,7 @@ export class SpearfishingSpotComponent extends TeamcraftPageComponent {
 
   public nodeData$: Observable<any>;
 
-  public bonuses$: Observable<any[]>;
+  public bonuses$: Observable<Object[]>;
 
   public links$: Observable<{ title: string, icon: string, url: string }[]>;
 
@@ -135,7 +135,9 @@ export class SpearfishingSpotComponent extends TeamcraftPageComponent {
                 });
               })
           );
-        }));
+        })).pipe(
+          map(bonuses => bonuses.flat())
+        );
       })
     );
 
