@@ -1,5 +1,4 @@
 import { app, BrowserWindow, ipcMain, protocol } from 'electron';
-import { createServer } from 'net';
 import { createServer as createHttpServer, Server } from 'http';
 import * as request from 'request';
 import { MainWindow } from './window/main-window';
@@ -21,7 +20,7 @@ export class TeamcraftDesktopApp {
   }
 
   start(): void {
-    app.releaseSingleInstanceLock();
+    // app.releaseSingleInstanceLock();
     app.setAsDefaultProtocolClient('teamcraft');
     let deepLink = this.store.get('deepLink', '');
 
@@ -44,18 +43,19 @@ export class TeamcraftDesktopApp {
         deepLink = '';
       }
 
-      this.isPortTaken(TeamcraftDesktopApp.MAIN_WINDOW_PORT).then(taken => {
-        if (taken) {
-          this.sendToAlreadyOpenedTC((this.argv[0] || '').replace('teamcraft://', ''));
-        } else {
+      request(`http://localhost:${TeamcraftDesktopApp.MAIN_WINDOW_PORT}${(this.argv[0] || '').replace('teamcraft://', '')}`, (err, res) => {
+        if (err) {
           this.bootApp();
+        } else {
+          (<any>app).isQuitting = true;
+          app.quit();
+          process.exit(0);
         }
       });
     });
 
     // Quit when all windows are closed.
     app.on('window-all-closed', () => {
-
       // On macOS specific close process
       if (process.platform !== 'darwin') {
         (<any>app).isQuitting = true;
@@ -73,30 +73,14 @@ export class TeamcraftDesktopApp {
     this.mainWindow.closed$.subscribe(() => {
       this.store.set('router:uri', deepLink);
     });
-  }
 
-  private isPortTaken(port: number): Promise<boolean> {
-    return new Promise((resolve, reject) => {
-      const tester = createServer()
-        .once('error', (err) => {
-          if ((<any>err).code !== 'EADDRINUSE') return reject(err);
-          resolve(true);
-        })
-        .once('listening', () => {
-          tester.once('close', () => {
-            resolve(false);
-          })
-            .close();
-        })
-        .listen(port, 'localhost');
+    app.on('before-quit', () => {
+      if (this.httpServer) {
+        this.httpServer.close(() => {
+          process.exit(0);
+        });
+      }
     });
-  }
-
-  private sendToAlreadyOpenedTC(url: string): void {
-    request(`http://localhost:${TeamcraftDesktopApp.MAIN_WINDOW_PORT}${url}`);
-    (<any>app).isQuitting = true;
-    app.quit();
-    process.exit(0);
   }
 
   private bootApp(): void {
@@ -112,7 +96,7 @@ export class TeamcraftDesktopApp {
       }
     });
 
-    loaderWindow.once('show', () => {
+    loaderWindow.once('ready-to-show', () => {
       this.mainWindow.createWindow();
       this.tray.createTray();
       this.httpServer = createHttpServer((req, res) => {
@@ -127,7 +111,9 @@ export class TeamcraftDesktopApp {
         }
         this.mainWindow.win.focus();
         this.mainWindow.win.show();
+        console.log(req.url);
         if (req.url.length > 1) {
+          console.log('SENDING URL TO MAIN');
           this.mainWindow.win.webContents.send('navigate', req.url);
         }
         res.writeHead(200);
@@ -141,6 +127,10 @@ export class TeamcraftDesktopApp {
         loaderWindow.hide();
         loaderWindow.close();
         this.mainWindow.show();
+        setTimeout(() => {
+          this.mainWindow.win.focus();
+          this.mainWindow.win.webContents.send('displayed', true);
+        }, 200);
       });
     });
 
