@@ -56,7 +56,9 @@ const patrons = [
   'Syl Varil',
   'Late Night Friendos',
   'The Hoe Depot',
-  '[Ultros] Nyumi Aitken'
+  '[Ultros] Nyumi Aitken',
+  'Syleboris Whitemaw (Phoenix)',
+  'Arimathean Industries™'
 ];
 
 const gifs = [
